pyproject.toml
[tool.poetry]
name = "base"
version = "0.1.0"
description = ""
authors = ["code406 <davidpm99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^4.2.9"
djangorestframework = "^3.14.0"
drf-keycloak-auth = "^0.3.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.13"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.7"}
djangorestframework-stubs = "^3.14.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "base.settings"
